refactor(agentspan): extract GraphStructureCompiler from AgentCompiler
Move the LangGraph-to-Conductor sub-compiler (compileGraphStructure,
buildLlmNodeTasks, buildSubgraphNodeTasks, findJoinPoint, allocRef,
buildForkMergeTask) into its own package-private class, mirroring
MultiAgentCompiler's existing (AgentCompiler) constructor pattern.
findJoinPoint/allocRef/buildForkMergeTask tighten from package-private
to private — nothing else called them. WORKFLOW_INPUTS widens from
private to package-private (also used by createWorkflow and
compileFrameworkPassthrough). applyRetryPolicy stays on AgentCompiler
(test-pinned) and is called qualified. No test files touched.
